Rename namespace from suppression to suppressions in core. (#3969)
