+++ src/Core/Suppressions/IFilterableIssue.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarLint.VisualStudio.Core.Suppression
+namespace SonarLint.VisualStudio.Core.Suppressions
 {
     /// <summary>
     /// Describes a single issue with the properties required for
